Split orchestrator workflow specs into focused test modules.
